✨ feat: 좋아요 여부 상태 유지 기능 구현
