+++ package.json
@@ -29,6 +29,7 @@
     "react-router": "^7.0.2",
     "react-router-dom": "^7.0.2",
     "react-secure-storage": "^1.3.2",
+    "react-toastify": "^11.0.0",
     "tailwind-merge": "^2.5.5",
     "zustand": "^5.0.2"
   },
