+++ src/pages/CourseContent.tsx
@@ -1,11 +1,11 @@
 import { useParams } from "react-router";
-import images from "../assets/images/importImages";
 import CourseContentDoc from "../components/main/courseContent/CourseContentDoc";
 import { useQuery } from "@tanstack/react-query";
 import { getCourses } from "../api/react-query/courseApi";
 import Loader from "../components/skeletonUI/Loader";
 import { useEffect } from "react";
 import { useCommentStore } from "../stores/main/comment/useCommentStore";
+import LikeButton from "../components/main/courseContent/LikeButton";
 
 export default function CourseContent() {
   const { contentId } = useParams<{ contentId: string }>();
@@ -26,6 +26,7 @@ export default function CourseContent() {
       setComments(courseData.comments);
     }
   }, [courseData, setComments]);
+
   return (
     <>
       {isCourseLoading ? (
@@ -43,14 +44,7 @@ export default function CourseContent() {
           </div>
           <div className="absolute bottom-0 left-0 w-full h-full top-[419px]">
             <div className="py-[38px] bg-[#F9FBFE] rounded-t-[40px]  shadow-[0_-8px_10px_0_rgba(48,72,100,0.25)] h-auto min-h-[1800px]">
-              <div className="absolute h-[49px] w-[49px] bg-custom-black opacity-80 rounded-full right-[60px] top-[-79px] flex justify-center items-center cursor-pointer">
-                <img
-                  src={images.white_heart_filled_icon}
-                  alt=""
-                  className="w-[25px] h-[21px]"
-                />
-              </div>
-
+              <LikeButton courseObj={courseData} />
               <div className="px-[61px] h-auto overflow-y-auto ">
                 <CourseContentDoc courseObj={courseData!} />
               </div>
